src/test/javascript/spec/app/entities/inscription-evenement/inscription-evenement.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { InscriptionEvenementService } from 'app/entities/inscription-evenement/inscription-evenement.service';
import { IInscriptionEvenement, InscriptionEvenement } from 'app/shared/model/inscription-evenement.model';

describe('Service Tests', () => {
  describe('InscriptionEvenement Service', () => {
    let injector: TestBed;
    let service: InscriptionEvenementService;
    let httpMock: HttpTestingController;
    let elemDefault: IInscriptionEvenement;
    let expectedResult: IInscriptionEvenement | IInscriptionEvenement[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(InscriptionEvenementService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new InscriptionEvenement(0, 'AAAAAAA', 'AAAAAAA');
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a InscriptionEvenement', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new InscriptionEvenement()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a InscriptionEvenement', () => {
        const returnedFromService = Object.assign(
          {
            loginParticipant: 'BBBBBB',
            passwordParticipant: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of InscriptionEvenement', () => {
        const returnedFromService = Object.assign(
          {
            loginParticipant: 'BBBBBB',
            passwordParticipant: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a InscriptionEvenement', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
